move ensure_plans to corehq.apps.accounting.bootstrap.utils
